# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/dashboard/buildings/{buildingId}/iot/devices": {
      "get": {
        "description": "Returns a normalized active device catalogue for Dashboard Page 07. Supports full-building catalogue (omitting floorId) and approved floor-scoped queries (floorId=4, 6). Zero database persistence.",
        "operationId": "DashboardIotController_getCatalogue",
        "parameters": [
          {
            "name": "buildingId",
            "required": true,
            "in": "path",
            "description": "Building identifier (currently E only)",
            "schema": {
              "example": "E",
              "type": "string"
            }
          },
          {
            "name": "floorId",
            "required": false,
            "in": "query",
            "description": "Application floor ID (4, 6). When omitted, returns full unfiltered catalogue.",
            "schema": {
              "example": "4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device catalogue returned successfully with live provenance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardDeviceCatalogueResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid buildingId, unsupported floor, or floor G"
          },
          "502": {
            "description": "Upstream IoT service error, timeout, or malformed payload"
          },
          "503": {
            "description": "IoT integration disabled, fixture mode, or missing credentials"
          }
        },
        "summary": "Get Dashboard IoT device catalogue (read-only, live provenance)",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/v1/dashboard/buildings/{buildingId}/iot/devices/{deviceId}/telemetry": {
      "get": {
        "description": "Returns normalized telemetry readings and coherent latest sample for a supported device (solar or avc). Strictly read-only and in-memory.",
        "operationId": "DashboardIotController_getTelemetry",
        "parameters": [
          {
            "name": "buildingId",
            "required": true,
            "in": "path",
            "description": "Building identifier (currently E only)",
            "schema": {
              "example": "E",
              "type": "string"
            }
          },
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "description": "Device EUI / ID",
            "schema": {
              "example": "70B3D57ED0073E9D",
              "type": "string"
            }
          },
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Query window start as ISO-8601 UTC string",
            "schema": {
              "example": "2026-09-23T12:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "stop",
            "required": true,
            "in": "query",
            "description": "Query window stop as ISO-8601 UTC string",
            "schema": {
              "example": "2026-09-26T12:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum readings to return (1..1000). Defaults to 1000.",
            "schema": {
              "default": 1000,
              "example": 1000,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Telemetry readings returned successfully with live provenance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardDeviceTelemetryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid buildingId, deviceId, range duration (>7d), limit (>1000), or unsupported device type"
          },
          "404": {
            "description": "Device not found"
          },
          "502": {
            "description": "Upstream IoT service error, timeout, or malformed payload"
          },
          "503": {
            "description": "IoT integration disabled, fixture mode, or missing credentials"
          }
        },
        "summary": "Get Dashboard IoT device telemetry (read-only, live provenance)",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
    "/api/v1/dashboard/buildings/{buildingId}/environment/sources": {
      "get": {
        "description": "Returns a list of registered Solar environment candidate sources for Building E. Reuses existing catalogue service and floor mapping. Strictly read-only and in-memory.",
        "operationId": "DashboardEnvironmentController_getSources",
        "parameters": [
          {
            "name": "buildingId",
            "required": true,
            "in": "path",
            "description": "Building identifier (currently E only)",
            "schema": {
              "example": "E",
              "type": "string"
            }
          },
          {
            "name": "floorId",
            "required": false,
            "in": "query",
            "description": "Application floor ID (4, 6). When omitted, returns full unfiltered catalogue of Solar sources.",
            "schema": {
              "example": "4",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Solar environment candidate sources list returned successfully with live provenance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardEnvironmentSourceListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid buildingId, unsupported floor, or floor G"
          },
          "502": {
            "description": "Upstream IoT service error, timeout, or malformed payload"
          },
          "503": {
            "description": "IoT integration disabled, fixture mode, or missing credentials"
          }
        },
        "summary": "Get Dashboard Solar environment sources list (read-only, live provenance)",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/v1/dashboard/buildings/{buildingId}/environment/summary": {
      "get": {
        "description": "Calculates request-scoped population summary (mean, min, max) of latest valid raw samples across candidate Solar sources within a requested window (max 24h). Max 20 sources, concurrency 2. Strictly in-memory, zero database writes.",
        "operationId": "DashboardEnvironmentController_getSummary",
        "parameters": [
          {
            "name": "buildingId",
            "required": true,
            "in": "path",
            "description": "Building identifier (currently E only)",
            "schema": {
              "example": "E",
              "type": "string"
            }
          },
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Inclusive query range start in ISO-8601 UTC format. Must be strictly before stop. Max duration 24 hours.",
            "schema": {
              "example": "2026-09-26T12:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "stop",
            "required": true,
            "in": "query",
            "description": "Inclusive query range stop in ISO-8601 UTC format. Must be strictly after start. Max duration 24 hours.",
            "schema": {
              "example": "2026-09-27T12:00:00.000Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Environment summary returned successfully with derived provenance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardEnvironmentSummaryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid buildingId, missing/invalid start/stop dates, or range duration > 24 hours"
          },
          "502": {
            "description": "Upstream IoT service error, timeout, or malformed payload"
          },
          "503": {
            "description": "IoT integration disabled, fixture mode, or missing credentials"
          }
        },
        "summary": "Get Dashboard bounded in-memory latest summary for Solar environment sources",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/v1/dashboard/buildings/{buildingId}/environment/sources/{deviceId}/readings": {
      "get": {
        "description": "Returns normalized readings, latest sample, coverage, and caveats for a single selected Solar source. Enforces server-authoritative solar type. Range duration cap 7 days, limit cap 1000.",
        "operationId": "DashboardEnvironmentController_getReadings",
        "parameters": [
          {
            "name": "buildingId",
            "required": true,
            "in": "path",
            "description": "Building identifier (currently E only)",
            "schema": {
              "example": "E",
              "type": "string"
            }
          },
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "description": "Device EUI / ID of Solar source",
            "schema": {
              "example": "8cf95720000a0123",
              "type": "string"
            }
          },
          {
            "name": "start",
            "required": true,
            "in": "query",
            "description": "Inclusive query range start in ISO-8601 UTC format. Must be strictly before stop.",
            "schema": {
              "example": "2026-09-26T12:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "stop",
            "required": true,
            "in": "query",
            "description": "Inclusive query range stop in ISO-8601 UTC format. Must be strictly after start.",
            "schema": {
              "example": "2026-09-27T12:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of readings to return (1..1000). Default is 1000.",
            "schema": {
              "default": 1000,
              "example": 1000,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Solar source environment readings returned successfully with live provenance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardEnvironmentReadingsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid buildingId, deviceId, range duration (>7d), limit (>1000), or device is not a Solar device"
          },
          "404": {
            "description": "Device not found"
          },
          "502": {
            "description": "Upstream IoT service error, timeout, or malformed payload"
          },
          "503": {
            "description": "IoT integration disabled, fixture mode, or missing credentials"
          }
        },
        "summary": "Get Dashboard selected Solar source raw environment readings (read-only, live provenance)",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
  },
  "info": {
    "title": "GIS-UIT Building E Digital Twin - Backend API",
    "description": "Internal API for device catalogue, marker positions, and persistence in Building E.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "DashboardCatalogueProvenanceDto": {
        "type": "object",
        "properties": {
          "mode": {
            "type": "string",
            "example": "live",
            "enum": [
              "live"
            ]
          },
          "sourceId": {
            "type": "string",
            "example": "iot-device-catalogue"
          },
          "sourceType": {
            "type": "string",
            "example": "iot_backend"
          },
          "fetchedAt": {
            "type": "string",
            "example": "2026-09-26T02:45:00.000Z",
            "description": "Application backend fetch timestamp (ISO-8601)"
          },
          "caveats": {
            "example": [
              "Dữ liệu development fallback được áp dụng cho tầng 4"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "mode",
          "sourceId",
          "sourceType",
          "fetchedAt"
        ]
      },
      "DashboardCatalogueMappingDto": {
        "type": "object",
        "properties": {
          "floorMode": {
            "type": "string",
            "example": "TEST_CURRENT_FLOOR_4_6_V1"
          },
          "developmentFallbackApplied": {
            "type": "boolean",
            "example": false,
            "description": "True when development fallback from floor 0 was applied"
          },
          "requestedUpstreamFloorLevel": {
            "type": "number",
            "example": 4,
            "nullable": true,
            "description": "Mapped upstream floor level requested, or null if full catalogue"
          }
        },
        "required": [
          "floorMode",
          "developmentFallbackApplied",
          "requestedUpstreamFloorLevel"
        ]
      },
      "DashboardCatalogueSummaryDto": {
        "type": "object",
        "properties": {
          "receivedCount": {
            "type": "number",
            "example": 25,
            "description": "Raw records received from upstream"
          },
          "acceptedCount": {
            "type": "number",
            "example": 25,
            "description": "Valid records accepted into catalogue"
          },
          "skippedCount": {
            "type": "number",
            "example": 0,
            "description": "Malformed records skipped"
          },
          "duplicateCount": {
            "type": "number",
            "example": 0,
            "description": "Duplicate device IDs skipped"
          },
          "truncated": {
            "type": "boolean",
            "example": false,
            "nullable": true,
            "description": "Upstream truncation indicator"
          }
        },
        "required": [
          "receivedCount",
          "acceptedCount",
          "skippedCount",
          "duplicateCount",
          "truncated"
        ]
      },
...
      "DashboardDeviceCatalogueItemDto": {
        "type": "object",
        "properties": {
          "externalDeviceId": {
            "type": "string",
            "example": "70B3D57ED0073E9D",
            "description": "Opaque upstream device identifier"
          },
          "sourceDeviceType": {
            "type": "string",
            "example": "solar",
            "description": "Raw source device type string"
          },
          "category": {
            "type": "string",
            "example": "solar",
            "description": "Mapped application device category"
          },
          "active": {
            "type": "boolean",
            "example": true,
            "description": "Catalogue registration/activity metadata (NOT an online/health signal)"
          },
          "sourceCreatedAt": {
            "type": "string",
            "example": "2026-09-20T19:09:17.355Z",
            "description": "Source record creation timestamp"
          },
          "sourceUpdatedAt": {
            "type": "string",
            "example": "2026-09-20T19:09:17.355Z",
            "description": "Source metadata update timestamp (NOT telemetry last-seen)"
          },
          "sourceLocation": {
            "$ref": "#/components/schemas/DashboardDeviceSourceLocationDto"
          },
          "displayFloorId": {
            "type": "string",
            "example": "4",
            "nullable": true,
            "description": "Viewer display floor ID, or null if unmapped"
          },
          "floorAssignment": {
            "type": "string",
            "example": "source",
            "enum": [
              "source",
              "development-fallback",
              "unmapped"
            ],
            "description": "Floor assignment provenance: source, development-fallback, or unmapped"
          }
        },
        "required": [
          "externalDeviceId",
          "sourceDeviceType",
          "category",
          "active",
          "sourceCreatedAt",
          "sourceUpdatedAt",
          "sourceLocation",
          "displayFloorId",
          "floorAssignment"
        ]
      },
      "DashboardDeviceCatalogueResponseDto": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "number",
            "example": 1
          },
          "buildingId": {
            "type": "string",
            "example": "E"
          },
          "requestedFloorId": {
            "type": "string",
            "example": "4",
            "nullable": true
          },
          "availability": {
            "type": "string",
            "example": "ready",
            "enum": [
              "ready",
              "empty"
            ]
          },
          "provenance": {
            "$ref": "#/components/schemas/DashboardCatalogueProvenanceDto"
          },
          "mapping": {
            "$ref": "#/components/schemas/DashboardCatalogueMappingDto"
          },
          "summary": {
            "$ref": "#/components/schemas/DashboardCatalogueSummaryDto"
          },
          "devices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardDeviceCatalogueItemDto"
            }
          }
        },
        "required": [
          "schemaVersion",
          "buildingId",
          "requestedFloorId",
          "availability",
          "provenance",
          "mapping",
          "summary",
          "devices"
        ]
      },
...
      "DashboardTelemetryQueryRangeDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "example": "2026-09-23T12:00:00.000Z"
          },
          "stop": {
            "type": "string",
            "example": "2026-09-26T12:00:00.000Z"
          },
          "limit": {
            "type": "number",
            "example": 1000
          }
        },
        "required": [
          "start",
          "stop",
          "limit"
        ]
      },
...
      "DashboardDeviceTelemetryResponseDto": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "number",
            "example": 1
          },
          "buildingId": {
            "type": "string",
            "example": "E"
          },
          "deviceId": {
            "type": "string",
            "example": "70B3D57ED0073E9D"
          },
          "deviceType": {
            "type": "string",
            "enum": [
              "solar",
              "avc"
            ],
            "example": "solar"
          },
          "availability": {
            "type": "string",
            "enum": [
              "ready",
              "empty"
            ],
            "example": "ready"
          },
          "provenance": {
            "$ref": "#/components/schemas/DashboardTelemetryProvenanceDto"
          },
          "queryRange": {
            "$ref": "#/components/schemas/DashboardTelemetryQueryRangeDto"
          },
          "coverage": {
            "type": "object"
          },
          "latestSample": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DashboardTelemetryLatestSampleDto"
              }
            ]
          },
          "telemetry": {
            "type": "object"
          }
        },
        "required": [
          "schemaVersion",
          "buildingId",
          "deviceId",
          "deviceType",
          "availability",
          "provenance",
          "queryRange",
          "coverage",
          "latestSample",
          "telemetry"
        ]
      },
...
      "DashboardEnvironmentSourceListResponseDto": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "number",
            "example": 1
          },
          "buildingId": {
            "type": "string",
            "example": "E"
          },
          "requestedFloorId": {
            "type": "string",
            "example": "4",
            "nullable": true
          },
          "availability": {
            "type": "string",
            "example": "ready",
            "enum": [
              "ready",
              "empty"
            ]
          },
          "provenance": {
            "$ref": "#/components/schemas/DashboardEnvironmentProvenanceDto"
          },
          "summary": {
            "$ref": "#/components/schemas/DashboardEnvironmentSourceSummaryDto"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardEnvironmentSourceItemDto"
            }
          }
        },
        "required": [
          "schemaVersion",
          "buildingId",
          "requestedFloorId",
          "availability",
          "provenance",
          "summary",
          "sources"
        ]
      },
...
      "DashboardEnvironmentSummaryResponseDto": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "number",
            "example": 1
          },
          "buildingId": {
            "type": "string",
            "example": "E"
          },
          "availability": {
            "type": "string",
            "example": "ready",
            "enum": [
              "ready",
              "partial",
              "empty"
            ]
          },
          "queryRange": {
            "$ref": "#/components/schemas/DashboardEnvironmentSummaryQueryRangeDto"
          },
          "provenance": {
            "$ref": "#/components/schemas/DashboardEnvironmentSummaryProvenanceDto"
          },
          "coverage": {
            "$ref": "#/components/schemas/DashboardEnvironmentSummaryCoverageDto"
          },
          "metrics": {
            "$ref": "#/components/schemas/DashboardEnvironmentSummaryMetricsDto"
          },
          "latestObservedAt": {
            "type": "string",
            "example": "2026-09-27T11:58:30.000Z",
            "nullable": true
          },
          "sourceResults": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardEnvironmentSourceResultDto"
            }
          }
        },
        "required": [
          "schemaVersion",
          "buildingId",
          "availability",
          "queryRange",
          "provenance",
          "coverage",
          "metrics",
          "latestObservedAt",
          "sourceResults"
        ]
      },
...
      "DashboardEnvironmentReadingsResponseDto": {
        "type": "object",
        "properties": {
          "schemaVersion": {
            "type": "number",
            "example": 1
          },
          "buildingId": {
            "type": "string",
            "example": "E"
          },
          "sourceId": {
            "type": "string",
            "example": "8cf95720000a0123"
          },
          "sourceDeviceType": {
            "type": "string",
            "example": "solar",
            "enum": [
              "solar"
            ]
          },
          "availability": {
            "type": "string",
            "example": "ready",
            "enum": [
              "ready",
              "empty"
            ]
          },
          "queryRange": {
            "type": "object",
            "example": {
              "start": "2026-09-26T12:00:00.000Z",
              "stop": "2026-09-27T12:00:00.000Z",
              "limit": 1000
            }
          },
          "provenance": {
            "$ref": "#/components/schemas/DashboardEnvironmentReadingsProvenanceDto"
          },
          "coverage": {
            "type": "object"
          },
          "latestSample": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DashboardEnvironmentLatestSampleDto"
              }
            ]
          },
          "readings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardEnvironmentReadingItemDto"
            }
          }
        },
        "required": [
          "schemaVersion",
          "buildingId",
          "sourceId",
          "sourceDeviceType",
          "availability",
          "queryRange",
          "provenance",
          "coverage",
          "latestSample",
          "readings"
        ]
      },
...
    }
  }
}